[feat] complete Stage 5
New option on main menu to print the status of the machine instead of printing it after each action.
New option on buy menu to back out of the menu without buying a drink.
Returns to main menu after buying a drink, whether successful or not.
New inner class to represent the drinks, with three inner classes of its own for the drinks.